requirements.txt
pyzbar
requests
pyotp
pillow